requirements.txt
opencv-python
dlib
numpy
scipy
PyYAML>=6.0
psutil
pygame